+++ src/constants/message.ts
@@ -5,6 +5,7 @@ const CART_MESSAGE = {
 
 const USER_MESSAGE = {
   FAIL_LOGIN: '아이디 혹은 비밀번호를 확인해주세요.',
+  FAIL_SIGNUP: '기입하신 정보를 확인해주세요.',
   ASK_WITH_DRAW: '탈퇴하시겠습니까?',
 };
